requirements.txt
asn1crypto==0.24.0
backports.csv==1.0.7
cffi==1.12.2
cryptography==2.6.1
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.0.7
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.2.0
django-pure-pagination==0.3.0
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.12.1
jdcal==1.4
odfpy==1.4.0
openpyxl==2.6.2
Pillow
pycparser==2.19
PyMySQL==0.9.1
pytz==2019.1
PyYAML==5.1
six==1.12.0
tablib==0.13.0
xlrd==1.2.0
xlwt==1.3.0
